refactor(server): split session.rs into focused submodules (TD-039)
Behavior-preserving decomposition of the 1976-line session.rs monolith into
four submodules: persistence (SessionRegistry, purge, OAuth), lifecycle
(create/restore/get_or_create), inference_setup (client construction, queue
init, initial save), and ticks (all background tasks including the Tier-2
simulation tick from #1198). session.rs becomes a thin hub re-exporting the
public API unchanged. Closes #1200 (partial).
