+++ parish/crates/parish-server/src/session.rs
@@ -4,41 +4,55 @@
 //! a `parish_sid` cookie.  Sessions survive server restarts because they are
 //! persisted in `saves/sessions.db` and their game state lives in
 //! `saves/<session_id>/parish_NNN.db`.
+//!
+//! # Module layout
+//!
+//! | Submodule             | Responsibility                                              |
+//! |-----------------------|-------------------------------------------------------------|
+//! | `persistence`         | [`SessionRegistry`], DB CRUD, OAuth linking, purge          |
+//! | `lifecycle`           | `get_or_create_session`, create, restore, finalize          |
+//! | `inference_setup`     | client construction, queue init, initial save               |
+//! | `ticks`               | background tick tasks (world, autosave, Tier-2, logs, …)   |
+//!
+//! The public API exported from this module is unchanged from the monolithic
+//! `session.rs` that preceded the split (TD-039, #1200).
 
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::sync::Arc;
-use std::sync::atomic::{AtomicU64, Ordering};
-use std::time::{Duration, Instant, SystemTime, UNIX_EPOCH};
+use std::sync::atomic::AtomicU64;
+use std::time::Duration;
 
-use dashmap::DashMap;
 use lru::LruCache;
 use tokio::sync::Mutex as TokioMutex;
 use tokio::task::JoinHandle;
 
 use parish_core::config::InferenceConfig;
 use parish_core::game_mod::{GameMod, PronunciationEntry};
-use parish_core::inference::{
-    AnyClient, InferenceQueue, InferenceWorkerConfig, spawn_inference_worker,
-};
 use parish_core::ipc::{GameConfig, ThemePalette};
-use parish_core::npc::manager::NpcManager;
 use parish_core::world::transport::TransportConfig;
-use parish_core::world::{DEFAULT_START_LOCATION, WorldState};
-
-use parish_core::event_bus::{EventBus as EventBusTrait, Topic};
 
 use parish_core::identity::IdentityStore;
 
-use crate::session_store_impl::{DbSessionStore, initialize_sessions_schema};
-use crate::state::{AppState, UiConfigSnapshot, build_app_state};
+use crate::state::{AppState, UiConfigSnapshot};
 
-// ── Public types ─────────────────────────────────────────────────────────────
+// ── Submodules ────────────────────────────────────────────────────────────────
 
-/// Re-export so the test in this module can reference the canonical constant
-/// without importing from parish_core directly.
+mod inference_setup;
+mod lifecycle;
+mod persistence;
+mod ticks;
+
+// ── Public re-exports ─────────────────────────────────────────────────────────
+
+pub use lifecycle::{CapacityExceededError, get_or_create_session};
+pub use persistence::{SessionRegistry, is_valid_session_id};
+// Re-export the autosave constant so the test in session/ticks.rs references
+// the canonical value without a separate parish_core import.
 pub use parish_core::AUTOSAVE_INTERVAL_SECS;
 
 // ── Idempotency cache ─────────────────────────────────────────────────────────
+
+use std::time::Instant;
 
 /// Default capacity of the idempotency LRU cache (process-wide).
 pub const IDEMPOTENCY_CACHE_CAPACITY: usize = 1000;
@@ -167,753 +181,7 @@ impl Drop for SessionEntry {
     }
 }
 
-// ── SessionRegistry ──────────────────────────────────────────────────────────
-
-/// In-memory session map backed by a SQLite persistence store.
-pub struct SessionRegistry {
-    sessions: DashMap<String, Arc<SessionEntry>>,
-    db: std::sync::Mutex<rusqlite::Connection>,
-    /// Running count of session-creation rejections since process start.
-    pub rejection_count: AtomicU64,
-}
-
-impl SessionRegistry {
-    /// Opens (or creates) `saves/sessions.db` and runs schema migrations.
-    pub fn open(saves_dir: &Path) -> rusqlite::Result<Self> {
-        let db_path = saves_dir.join("sessions.db");
-        let conn = rusqlite::Connection::open(&db_path)?;
-        initialize_sessions_schema(&conn)?;
-        Ok(Self {
-            sessions: DashMap::new(),
-            db: std::sync::Mutex::new(conn),
-            rejection_count: AtomicU64::new(0),
-        })
-    }
-
-    /// Returns the number of sessions currently held in memory.
-    pub fn active_count(&self) -> usize {
-        self.sessions.len()
-    }
-
-    /// Returns `true` when the number of live sessions is at or above `cap`.
-    ///
-    /// Callers should check this before creating a new session and return
-    /// `503 Service Unavailable` if it holds. Increments `rejection_count`
-    /// when at capacity so callers don't have to manage that separately.
-    pub fn is_at_capacity(&self, cap: usize) -> bool {
-        let current = self.active_count();
-        if current >= cap {
-            self.rejection_count.fetch_add(1, Ordering::Relaxed);
-            true
-        } else {
-            false
-        }
-    }
-
-    pub fn now_unix() -> u64 {
-        SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .map(|d| d.as_secs())
-            .unwrap_or(0)
-    }
-
-    fn now_iso() -> String {
-        chrono::Utc::now().to_rfc3339()
-    }
-
-    /// Returns `true` if `session_id` is recorded in sessions.db.
-    pub fn exists_in_db(&self, session_id: &str) -> bool {
-        let db = self.db.lock().unwrap();
-        db.query_row("SELECT 1 FROM sessions WHERE id = ?1", [session_id], |_| {
-            Ok(())
-        })
-        .is_ok()
-    }
-
-    /// Inserts a new session row into sessions.db.
-    pub fn persist_new(&self, session_id: &str) {
-        let now = Self::now_iso();
-        let db = self.db.lock().unwrap();
-        if let Err(e) = db.execute(
-            "INSERT OR IGNORE INTO sessions (id, created_at, last_active) VALUES (?1, ?2, ?2)",
-            rusqlite::params![session_id, now],
-        ) {
-            tracing::warn!(session_id = %session_id, error = %e, "persist_new failed");
-        }
-    }
-
-    /// Updates the `last_active` timestamp for a session in sessions.db.
-    pub fn update_last_active(&self, session_id: &str) {
-        let now = Self::now_iso();
-        let db = self.db.lock().unwrap();
-        if let Err(e) = db.execute(
-            "UPDATE sessions SET last_active = ?1 WHERE id = ?2",
-            rusqlite::params![now, session_id],
-        ) {
-            tracing::warn!(session_id = %session_id, error = %e, "update_last_active failed");
-        }
-    }
-
-    /// Returns the session_id linked to an OAuth identity, if any.
-    pub fn find_by_oauth(&self, provider: &str, provider_user_id: &str) -> Option<String> {
-        let db = self.db.lock().unwrap();
-        db.query_row(
-            "SELECT session_id FROM oauth_accounts
-             WHERE provider = ?1 AND provider_user_id = ?2",
-            rusqlite::params![provider, provider_user_id],
-            |row| row.get(0),
-        )
-        .ok()
-    }
-
-    /// Associates an OAuth identity with a session_id, storing the user's display name.
-    pub fn link_oauth(
-        &self,
-        provider: &str,
-        provider_user_id: &str,
-        session_id: &str,
-        display_name: &str,
-    ) {
-        let db = self.db.lock().unwrap();
-        match db.execute(
-            "INSERT OR REPLACE INTO oauth_accounts
-             (provider, provider_user_id, session_id, display_name) VALUES (?1, ?2, ?3, ?4)",
-            rusqlite::params![provider, provider_user_id, session_id, display_name],
-        ) {
-            Ok(rows) => tracing::info!(
-                provider = %provider,
-                provider_user_id = %provider_user_id,
-                session_id = %session_id,
-                display_name = %display_name,
-                rows = rows,
-                "link_oauth stored account"
-            ),
-            Err(e) => tracing::error!(
-                provider = %provider,
-                provider_user_id = %provider_user_id,
-                session_id = %session_id,
-                error = %e,
-                "link_oauth DB write failed"
-            ),
-        }
-    }
-
-    /// Returns a session from the in-memory map.
-    pub fn get_in_memory(&self, session_id: &str) -> Option<Arc<SessionEntry>> {
-        self.sessions.get(session_id).map(|e| Arc::clone(&*e))
-    }
-
-    /// Inserts a session into the in-memory map.
-    pub fn insert(&self, session_id: String, entry: Arc<SessionEntry>) {
-        self.sessions.insert(session_id, entry);
-    }
-
-    /// Removes sessions that have been idle longer than `max_age`.
-    ///
-    /// The sessions' background tick tasks are implicitly cancelled when
-    /// their `JoinHandle`s are dropped via the evicted `SessionEntry`.
-    pub fn cleanup_stale(&self, max_age: Duration) {
-        let cutoff = Self::now_unix().saturating_sub(max_age.as_secs());
-        self.sessions
-            .retain(|_, entry| entry.last_active.load(Ordering::Relaxed) >= cutoff);
-    }
-
-    /// Purges sessions abandoned for longer than `max_age` from disk
-    /// (sessions.db row + saves/<session_id>/ directory).
-    ///
-    /// Distinct from [`cleanup_stale`]: that one only clears the
-    /// in-memory map. Disk state is what needs removing here (#482) —
-    /// otherwise long-running deployments accumulate dead sessions
-    /// forever. `max_age` is expected to be much longer than the
-    /// in-memory TTL (e.g. 30 days vs 2 hours) so users can still
-    /// restore a session from the cookie on their next visit for
-    /// reasonable idle windows.
-    ///
-    /// Returns the number of sessions purged, so the caller can log
-    /// the scope of the sweep.
-    pub fn purge_expired_disk_sessions(&self, saves_root: &Path, max_age: Duration) -> usize {
-        let cutoff_secs = Self::now_unix().saturating_sub(max_age.as_secs());
-        let cutoff = match chrono::DateTime::<chrono::Utc>::from_timestamp(cutoff_secs as i64, 0) {
-            Some(dt) => dt.to_rfc3339(),
-            None => {
-                tracing::warn!(
-                    cutoff_secs = cutoff_secs,
-                    "purge_expired_disk_sessions: cutoff timestamp out of range, skipping sweep"
-                );
-                return 0;
-            }
-        };
-
-        // Find expired session ids + drop their sessions.db rows in a
-        // single transaction so the filesystem cleanup below can't get
-        // out of sync with the DB if the process dies mid-sweep.
-        let expired_ids: Vec<String> = {
-            let db = match self.db.lock() {
-                Ok(g) => g,
-                Err(poisoned) => poisoned.into_inner(),
-            };
-            let mut collected = Vec::new();
-            let select_result = (|| -> rusqlite::Result<()> {
-                let mut stmt = db.prepare("SELECT id FROM sessions WHERE last_active < ?1")?;
-                let mut rows = stmt.query([&cutoff])?;
-                while let Some(row) = rows.next()? {
-                    collected.push(row.get::<_, String>(0)?);
-                }
-                Ok(())
-            })();
-            if let Err(e) = select_result {
-                tracing::warn!(error = %e, "purge_expired_disk_sessions: DB read failed");
-                return 0;
-            }
-            // Drop rows for the ids we collected inside an explicit
-            // transaction.  Both DELETEs must commit atomically: if the
-            // process crashes between them, oauth_accounts rows would be
-            // left pointing at a non-existent session_id, letting the
-            // next login for that OAuth identity silently resurrect a
-            // ghost session (#593, #482).
-            //
-            // Invariant: DB rows are deleted *before* filesystem cleanup
-            // (see below).  A residual saves/<id>/ directory with no DB
-            // row is harmless; an oauth_accounts row pointing at a missing
-            // sessions row is not.
-            if !collected.is_empty() {
-                let tx_result = (|| -> rusqlite::Result<()> {
-                    let tx = db.unchecked_transaction()?;
-                    let placeholders = vec!["?"; collected.len()].join(",");
-                    let params: Vec<&dyn rusqlite::ToSql> = collected
-                        .iter()
-                        .map(|s| s as &dyn rusqlite::ToSql)
-                        .collect();
-                    let sql = format!("DELETE FROM sessions WHERE id IN ({placeholders})");
-                    tx.execute(&sql, params.as_slice())?;
-                    // Also drop oauth links for those sessions — otherwise
-                    // the next login for the same provider_user_id would
-                    // resurrect a dead session_id. (#482 sibling concern.)
-                    let oauth_sql =
-                        format!("DELETE FROM oauth_accounts WHERE session_id IN ({placeholders})");
-                    tx.execute(&oauth_sql, params.as_slice())?;
-                    tx.commit()
-                })();
-                if let Err(e) = tx_result {
-                    tracing::warn!(error = %e, "purge_expired_disk_sessions: DB delete failed");
-                    return 0;
-                }
-            }
-            collected
-        };
-
-        if expired_ids.is_empty() {
-            return 0;
-        }
-
-        // Best-effort filesystem cleanup. A failure here is logged but
-        // doesn't undo the DB delete — a residual saves/<id>/ directory
-        // with no DB row is harmless (eventually reaped by OS-level
-        // cleanup or a later sweep once we have directory-age scanning).
-        //
-        // #595 — Validate each session ID before building a path so that a
-        // corrupted or tampered DB row cannot cause remove_dir_all to delete
-        // directories outside the saves root.  Two layers of defence:
-        //   1. Allowlist check: the ID must consist only of lowercase hex
-        //      digits and hyphens (UUID v4 format).  Anything else — including
-        //      `..`, `/`, `\`, or unusual chars — is rejected before we even
-        //      call Path::join.
-        //   2. Containment check: after joining, canonicalize the candidate
-        //      path and assert it starts with the canonicalized saves root.
-        //      This catches any edge case the regex might miss.
-        let canonical_saves_root = match saves_root.canonicalize() {
-            Ok(p) => p,
-            Err(e) => {
-                tracing::warn!(
-                    error = %e,
-                    "purge_expired_disk_sessions: cannot canonicalize saves_root, skipping fs cleanup"
-                );
-                return expired_ids.len();
-            }
-        };
-
-        for id in &expired_ids {
-            // Layer 1: allowlist — UUID v4 looks like
-            // `xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx` (hex + hyphens only).
-            // Uses the shared `is_valid_session_id` helper (same rule as
-            // cookie ingress in `get_or_create_session`).
-            if !is_valid_session_id(id) {
-                tracing::warn!(
-                    session_id = %id,
-                    "purge_expired_disk_sessions: rejected unsafe session ID, skipping fs remove"
-                );
-                continue;
-            }
-
-            let session_dir = saves_root.join(id);
-            if !session_dir.exists() {
-                continue;
-            }
-
-            // Layer 2: containment — canonicalize the resolved path and verify
-            // it stays inside the saves root (guards against symlink tricks or
-            // any bypass of the allowlist above).
-            let canonical_dir = match session_dir.canonicalize() {
-                Ok(p) => p,
-                Err(e) => {
-                    tracing::warn!(
-                        session_id = %id,
-                        path = %session_dir.display(),
-                        error = %e,
-                        "purge_expired_disk_sessions: cannot canonicalize session dir, skipping"
-                    );
-                    continue;
-                }
-            };
-            if !canonical_dir.starts_with(&canonical_saves_root) {
-                tracing::warn!(
-                    session_id = %id,
-                    path = %canonical_dir.display(),
-                    saves_root = %canonical_saves_root.display(),
-                    "purge_expired_disk_sessions: path escapes saves root, skipping fs remove"
-                );
-                continue;
-            }
-
-            match std::fs::remove_dir_all(&session_dir) {
-                Ok(()) => {
-                    tracing::info!(
-                        session_id = %id,
-                        path = %session_dir.display(),
-                        "purge_expired_disk_sessions: removed saves directory"
-                    );
-                }
-                Err(e) => {
-                    tracing::warn!(
-                        session_id = %id,
-                        path = %session_dir.display(),
-                        error = %e,
-                        "purge_expired_disk_sessions: failed to remove saves directory"
-                    );
-                }
-            }
-        }
-
-        expired_ids.len()
-    }
-}
-
-// ── Session ID validation ─────────────────────────────────────────────────────
-
-/// Returns `true` when `id` is a structurally valid session ID.
-///
-/// A valid ID contains only lowercase hex digits and hyphens (the UUID v4
-/// character set: `xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx`).  The check also
-/// rejects `..` to make it impossible to construct a path-traversal sequence
-/// before the value ever reaches `Path::join`.
-///
-/// This is the single source of truth for session-ID validation; both
-/// [`get_or_create_session`] (cookie ingress) and
-/// [`SessionRegistry::purge_expired_disk_sessions`] (DB-sourced IDs) call
-/// this helper.
-pub fn is_valid_session_id(id: &str) -> bool {
-    !id.is_empty() && !id.contains("..") && id.chars().all(|c| c.is_ascii_hexdigit() || c == '-')
-}
-
-// ── Public session resolution ─────────────────────────────────────────────────
-
-/// Error returned by [`get_or_create_session`] when the server is at capacity.
-///
-/// The middleware maps this to `503 Service Unavailable` with a
-/// `Retry-After: 30` header.  Existing sessions (returning visitors with a
-/// valid cookie) are never refused — capacity is only checked when a brand-new
-/// session would be created.
-#[derive(Debug)]
-pub struct CapacityExceededError {
-    pub current: usize,
-    pub cap: usize,
-}
-
-/// Returns the session for `cookie_id`, restoring or creating one as needed.
-///
-/// Returns `Ok((session_id, entry, is_new))` where `is_new` is `true` when a
-/// fresh `parish_sid` cookie must be set on the response.
-///
-/// Returns `Err(CapacityExceededError)` when `global.max_concurrent_sessions`
-/// is set and the server is already at capacity.  Only new-session creation is
-/// gated — returning visitors whose session is already in memory or can be
-/// restored from the DB are never rejected.
-pub async fn get_or_create_session(
-    global: &Arc<GlobalState>,
-    cookie_id: Option<&str>,
-) -> Result<(String, Arc<SessionEntry>, bool), CapacityExceededError> {
-    // 1. Hot path: session already in memory.
-    //    Reject malformed cookie values before any DB lookup or path join.
-    if let Some(id) = cookie_id {
-        if !is_valid_session_id(id) {
-            tracing::warn!(
-                cookie_value = %id,
-                "get_or_create_session: invalid session ID format, treating as no session"
-            );
-            // Fall through to step 3: create a fresh session.
-        } else {
-            // 1a. Hot path: session already in memory.
-            if let Some(entry) = global.sessions.get_in_memory(id) {
-                entry
-                    .last_active
-                    .store(SessionRegistry::now_unix(), Ordering::Relaxed);
-                global.sessions.update_last_active(id);
-                return Ok((id.to_string(), entry, false));
-            }
-            // 2. Session known in DB but evicted from memory — restore it.
-            if global.sessions.exists_in_db(id) {
-                match restore_session(global, id).await {
-                    Ok(entry) => {
-                        global.sessions.insert(id.to_string(), Arc::clone(&entry));
-                        global.sessions.update_last_active(id);
-                        return Ok((id.to_string(), entry, false));
-                    }
-                    Err(e) => {
-                        tracing::warn!(session_id = %id, "Session restore failed: {}. Starting fresh.", e);
-                    }
-                }
-            }
-        }
-    }
-
-    // 3. No usable session — admission control check before creating a new one.
-    if let Some(cap) = global.max_concurrent_sessions {
-        let current = global.sessions.active_count();
-        if global.sessions.is_at_capacity(cap) {
-            let rejection_count = global.sessions.rejection_count.load(Ordering::Relaxed);
-            tracing::info!(
-                current_session_count = current,
-                capacity = cap,
-                rejection_count,
-                "admission-control: session capacity exceeded, rejecting new session"
-            );
-            return Err(CapacityExceededError { current, cap });
-        }
-    }
-
-    // 4. Create a new session.
-    let session_id = uuid::Uuid::new_v4().to_string();
-    let entry = create_session(global, &session_id).await;
-    global.sessions.persist_new(&session_id);
-    global
-        .sessions
-        .insert(session_id.clone(), Arc::clone(&entry));
-    tracing::info!(
-        current_session_count = global.sessions.active_count(),
-        capacity = global.max_concurrent_sessions,
-        "new session created"
-    );
-    Ok((session_id, entry, true))
-}
-
-// ── Session creation ──────────────────────────────────────────────────────────
-
-async fn create_session(global: &Arc<GlobalState>, session_id: &str) -> Arc<SessionEntry> {
-    let session_saves = global.saves_dir.join(session_id);
-    std::fs::create_dir_all(&session_saves).ok();
-
-    let world_path = global.world_path.clone();
-    let data_dir = global.data_dir.clone();
-    let (world, npc_manager) = tokio::task::spawn_blocking(move || {
-        let world = WorldState::from_parish_file(&world_path, DEFAULT_START_LOCATION)
-            .unwrap_or_else(|e| {
-                tracing::warn!("Session init: failed to load world: {}. Using default.", e);
-                WorldState::new()
-            });
-        let mut npc_manager = NpcManager::load_from_file(&data_dir.join("npcs.json"))
-            .unwrap_or_else(|e| {
-                tracing::warn!("Session init: failed to load npcs.json: {}. No NPCs.", e);
-                NpcManager::new()
-            });
-        npc_manager.assign_tiers(&world, &[]);
-        (world, npc_manager)
-    })
-    .await
-    .expect("session init blocking task panicked");
-
-    let (client, config) = build_session_client(global);
-    let cloud_client = build_session_cloud_client(global);
-    let game_mod = global.game_mod.clone();
-
-    let flags_path = global.data_dir.join("parish-flags.json");
-    let session_store = Arc::new(DbSessionStore::new(session_saves.clone()));
-
-    let log_to_disk = parish_core::inference::file_log::resolve_enabled(
-        false,
-        global.inference_config.log_to_disk,
-    );
-    let inference_file_log = parish_core::inference::file_log::InferenceFileLog::spawn(
-        &session_saves,
-        log_to_disk,
-        Some(&config.base_url),
-    );
-    let chat_transcript_log = parish_core::chat_transcript::ChatTranscriptLog::spawn_with_flag(
-        &session_saves,
-        inference_file_log.session_id().to_string(),
-        inference_file_log.enabled_flag(),
-    );
-
-    let app_state = build_app_state(
-        session_id.to_string(),
-        world,
-        npc_manager,
-        client.clone(),
-        config,
-        cloud_client,
-        global.transport.clone(),
-        global.ui_config.clone(),
-        global.theme_palette.clone(),
-        session_saves.clone(),
-        global.data_dir.clone(),
-        game_mod,
-        flags_path,
-        global.inference_config.clone(), // (#417) propagate TOML-configured timeouts
-        session_store,
-        inference_file_log,
-        chat_transcript_log,
-    );
-
-    if let Err(e) = init_session_save(&app_state, &session_saves).await {
-        tracing::warn!("Session initial save failed: {}", e);
-    }
-
-    finalize_session_entry(app_state, client).await
-}
-
-/// Shared tail of session entry construction: starts the inference queue
-/// (if a client is configured), spawns background ticks, and returns the
-/// wrapped [`SessionEntry`].
-async fn finalize_session_entry(
-    app_state: Arc<AppState>,
-    client: Option<AnyClient>,
-) -> Arc<SessionEntry> {
-    if let Some(ref c) = client {
-        init_inference_queue(&app_state, c.clone()).await;
-    }
-
-    let shutdown_token = tokio_util::sync::CancellationToken::new();
-    let handles = spawn_session_ticks(Arc::clone(&app_state), shutdown_token.clone());
-
-    Arc::new(SessionEntry {
-        app_state,
-        last_active: AtomicU64::new(SessionRegistry::now_unix()),
-        _shutdown_token: shutdown_token,
-        _tick_handles: handles,
-    })
-}
-
-// ── Session restoration ───────────────────────────────────────────────────────
-
-async fn restore_session(
-    global: &Arc<GlobalState>,
-    session_id: &str,
-) -> Result<Arc<SessionEntry>, String> {
-    let session_saves = global.saves_dir.join(session_id);
-    if !session_saves.exists() {
-        return Err(format!(
-            "saves directory {} does not exist",
-            session_saves.display()
-        ));
-    }
-
-    // Select the most recently modified `.db` file.  In normal play there is
-    // only one save per session, but branching can create additional files.
-    // Using mtime rather than alphabetical order avoids restoring a stale
-    // branch when newer ones exist (#632).
-    let saves_for_scan = session_saves.clone();
-    let db_path = tokio::task::spawn_blocking(move || -> Result<PathBuf, String> {
-        let mut entries: Vec<(PathBuf, std::time::SystemTime)> = std::fs::read_dir(&saves_for_scan)
-            .map_err(|e| e.to_string())?
-            .filter_map(|e| e.ok())
-            .filter(|e| e.path().extension().is_some_and(|ext| ext == "db"))
-            .filter_map(|e| {
-                let meta = e.metadata().ok()?;
-                let mtime = meta.modified().ok()?;
-                Some((e.path(), mtime))
-            })
-            .collect();
-        entries.sort_by_key(|b| std::cmp::Reverse(b.1));
-        entries
-            .into_iter()
-            .next()
-            .map(|(p, _)| p)
-            .ok_or_else(|| "no save files found".to_string())
-    })
-    .await
-    .map_err(|e| e.to_string())??;
-
-    // Load snapshot from the first branch.
-    let db_path_clone = db_path.clone();
-    let (snapshot, branch_id, branch_name) = tokio::task::spawn_blocking(move || {
-        use parish_core::persistence::Database;
-        let db = Database::open(&db_path_clone).map_err(|e| e.to_string())?;
-        let branches = db.list_branches().map_err(|e| e.to_string())?;
-        let branch = branches.into_iter().next().ok_or("no branches")?;
-        let (_, snapshot) = db
-            .load_latest_snapshot(branch.id)
-            .map_err(|e| e.to_string())?
-            .ok_or("no snapshots")?;
-        Ok::<_, String>((snapshot, branch.id, branch.name))
-    })
-    .await
-    .map_err(|e| e.to_string())??;
-
-    // Load fresh static world data, then apply the saved snapshot.
-    let world_path = global.world_path.clone();
-    let data_dir = global.data_dir.clone();
-    let (mut world, mut npc_manager) = tokio::task::spawn_blocking(move || {
-        let world = WorldState::from_parish_file(&world_path, DEFAULT_START_LOCATION)
-            .unwrap_or_else(|_| WorldState::new());
-        let npc_manager = NpcManager::load_from_file(&data_dir.join("npcs.json"))
-            .unwrap_or_else(|_| NpcManager::new());
-        (world, npc_manager)
-    })
-    .await
-    .map_err(|e| e.to_string())?;
-
-    snapshot.restore(&mut world, &mut npc_manager);
-    npc_manager.assign_tiers(&world, &[]);
-
-    let (client, config) = build_session_client(global);
-    let cloud_client = build_session_cloud_client(global);
-    let game_mod = global.game_mod.clone();
-
-    let flags_path = global.data_dir.join("parish-flags.json");
-    let session_store = Arc::new(DbSessionStore::new(session_saves.clone()));
-
-    // Persistent inference + transcript logs for this session. Same
-    // session_id is embedded in both filenames so they pair on disk.
-    let log_to_disk = parish_core::inference::file_log::resolve_enabled(
-        false, // server has no --no-inference-log flag; env var wins
-        global.inference_config.log_to_disk,
-    );
-    let inference_file_log = parish_core::inference::file_log::InferenceFileLog::spawn(
-        &session_saves,
-        log_to_disk,
-        Some(&config.base_url),
-    );
-    let chat_transcript_log = parish_core::chat_transcript::ChatTranscriptLog::spawn_with_flag(
-        &session_saves,
-        inference_file_log.session_id().to_string(),
-        inference_file_log.enabled_flag(),
-    );
-
-    let app_state = build_app_state(
-        session_id.to_string(),
-        world,
-        npc_manager,
-        client.clone(),
-        config,
-        cloud_client,
-        global.transport.clone(),
-        global.ui_config.clone(),
-        global.theme_palette.clone(),
-        session_saves.clone(),
-        global.data_dir.clone(),
-        game_mod,
-        flags_path,
-        global.inference_config.clone(), // (#417) propagate TOML-configured timeouts
-        session_store,
-        inference_file_log,
-        chat_transcript_log,
-    );
-
-    if let Some(ref c) = client {
-        init_inference_queue(&app_state, c.clone()).await;
-    }
-
-    // Acquire advisory lock on the restored save file so another server
-    // instance (or a headless CLI) cannot concurrently write to it (#425).
-    // If a peer already holds the lock we log a warning and continue:
-    // refusing to start would leave the user with no session at all, and
-    // per-process ownership makes strict mutual exclusion across
-    // containers out of scope for this handler. The lock is stored on
-    // AppState.save_lock so it lives for the session's lifetime.
-    let locked = parish_core::persistence::SaveFileLock::try_acquire(&db_path);
-    if locked.is_none() {
-        tracing::warn!(
-            path = %db_path.display(),
-            session_id = %session_id,
-            "SaveFileLock::try_acquire returned None on session resume — save file appears in use by another instance",
-        );
-    }
-    *app_state.save_lock.lock().await = locked;
-    *app_state.save_path.lock().await = Some(db_path);
-    *app_state.current_branch_id.lock().await = Some(branch_id);
-    *app_state.current_branch_name.lock().await = Some(branch_name);
-
-    Ok(finalize_session_entry(app_state, client).await)
-}
-
-// ── Helpers ───────────────────────────────────────────────────────────────────
-
-async fn init_inference_queue(app_state: &Arc<AppState>, client: AnyClient) {
-    let (interactive_tx, interactive_rx) = tokio::sync::mpsc::channel(16);
-    let (background_tx, background_rx) = tokio::sync::mpsc::channel(32);
-    let (batch_tx, batch_rx) = tokio::sync::mpsc::channel(64);
-    let provider =
-        parish_core::config::Provider::from_str_loose(&app_state.config.lock().await.provider_name)
-            .unwrap_or_default();
-    let worker = spawn_inference_worker(
-        client,
-        InferenceWorkerConfig {
-            interactive_rx,
-            background_rx,
-            batch_rx,
-            log: app_state.inference_log.clone(),
-            file_log: app_state.inference_file_log.clone(),
-            provider,
-            timeout_config: app_state.inference_config.clone(),
-        },
-    );
-    let queue = InferenceQueue::new(interactive_tx, background_tx, batch_tx);
-    *app_state.inference_queue.lock().await = Some(queue);
-    *app_state.worker_handle.lock().await = Some(worker);
-}
-
-/// Saves the initial world snapshot into `saves/<session_id>/parish_001.db`.
-async fn init_session_save(app_state: &Arc<AppState>, session_saves: &Path) -> Result<(), String> {
-    use parish_core::persistence::Database;
-    use parish_core::persistence::picker::new_save_path;
-    use parish_core::persistence::snapshot::GameSnapshot;
-
-    let snapshot = {
-        let world = app_state.world.lock().await;
-        let npc_manager = app_state.npc_manager.lock().await;
-        GameSnapshot::capture(&world, &npc_manager)
-    };
-
-    let save_path = new_save_path(session_saves);
-    let save_path_clone = save_path.clone();
-
-    let branch_id = tokio::task::spawn_blocking(move || -> Result<i64, String> {
-        let db = Database::open(&save_path_clone).map_err(|e| e.to_string())?;
-        let branch_id = db.create_branch("main", None).map_err(|e| e.to_string())?;
-        db.save_snapshot(branch_id, &snapshot)
-            .map_err(|e| e.to_string())?;
-        Ok(branch_id)
-    })
-    .await
-    .map_err(|e| e.to_string())??;
-
-    // Advisory lock on the freshly-initialised save file so peer
-    // instances don't write to it concurrently (#425). For a just-created
-    // save we expect the lock to always succeed, but we stay defensive:
-    // warn if the lock fails rather than silently proceeding.
-    let locked = parish_core::persistence::SaveFileLock::try_acquire(&save_path);
-    if locked.is_none() {
-        tracing::warn!(
-            path = %save_path.display(),
-            "SaveFileLock::try_acquire returned None on init_session_save — new save file unexpectedly locked",
-        );
-    }
-    *app_state.save_lock.lock().await = locked;
-    *app_state.save_path.lock().await = Some(save_path);
-    *app_state.current_branch_id.lock().await = Some(branch_id);
-    *app_state.current_branch_name.lock().await = Some("main".to_string());
-
-    Ok(())
-}
+// ── Module-level constant (re-exported for other modules) ────────────────────
 
 /// Maximum number of gossip propagations performed on a single world tick.
 ///
@@ -922,1055 +190,4 @@ async fn init_session_save(app_state: &Arc<AppState>, session_saves: &Path) -> R
 /// tick, stalling the 5-second world tick visibly for all connected
 /// clients (#466). Budgeting keeps each tick cheap; remaining groups get
 /// picked up by the next tick via a round-robin cursor.
-const GOSSIP_BUDGET_PER_TICK: usize = 20;
-
-/// Spawns the three per-session background tasks and returns their handles.
-///
-/// Each task observes `shutdown_token` via `tokio::select!` so it exits
-/// cleanly when the token is cancelled (e.g. on session eviction) rather than
-/// running until its `JoinHandle` is aborted (#228).
-fn spawn_session_ticks(
-    state: Arc<AppState>,
-    shutdown_token: tokio_util::sync::CancellationToken,
-) -> Vec<JoinHandle<()>> {
-    let mut handles = Vec::with_capacity(4);
-
-    // ── Character-log subscriber ───────────────────────────────────────────
-    //
-    // Per-character markdown logs (rule #12: orchestration lives in
-    // parish-core; this is the thin server-side wiring). Gated by the
-    // `character-logs` flag (default on). Subscribes BEFORE the profile
-    // rewrite so no events fired during/just after initial profile
-    // generation are lost.
-    {
-        let s = Arc::clone(&state);
-        let token = shutdown_token.clone();
-        handles.push(tokio::spawn(async move {
-            use parish_core::character_log::{CharacterLogManager, FEATURE_FLAG};
-
-            let enabled = {
-                let cfg = s.config.lock().await;
-                !cfg.flags.is_disabled(FEATURE_FLAG)
-            };
-            if !enabled {
-                return;
-            }
-            let app_name = parish_core::game_mod::app_name_from_mod(&s.game_mod);
-            let mut current_branch = s.current_branch_id.lock().await.unwrap_or(1);
-            let mut manager = CharacterLogManager::new(&app_name, current_branch, true);
-            let mut rx = {
-                let world = s.world.lock().await;
-                world.event_bus.subscribe()
-            };
-            {
-                let world = s.world.lock().await;
-                let npc_mgr = s.npc_manager.lock().await;
-                if let Err(e) = manager.write_all_profiles(&world, &npc_mgr) {
-                    tracing::warn!(error = %e, "character-log profile write failed");
-                }
-            }
-            loop {
-                tokio::select! {
-                    _ = token.cancelled() => break,
-                    result = rx.recv() => match result {
-                        Ok(event) => {
-                            // Rebind manager when the active branch has changed
-                            // (e.g. load_branch / create_branch). Without this the
-                            // writer keeps appending to the original branch's
-                            // log directory after a branch switch (#1011).
-                            let bid = s.current_branch_id.lock().await.unwrap_or(1);
-                            if bid != current_branch {
-                                current_branch = bid;
-                                manager = CharacterLogManager::new(&app_name, bid, true);
-                                let world = s.world.lock().await;
-                                let npc_mgr = s.npc_manager.lock().await;
-                                if let Err(e) = manager.write_all_profiles(&world, &npc_mgr) {
-                                    tracing::warn!(error = %e, "character-log profile write failed after branch switch");
-                                }
-                            }
-                            // Clone the Arc<AppState> (cheap) and run the blocking
-                            // I/O task in a dedicated thread pool, avoiding lock
-                            // contention on the async side (#1012).
-                            let mgr_clone = manager.clone();
-                            let evt_clone = event.clone();
-                            let state_clone = Arc::clone(&s);
-                            let handle = tokio::task::spawn_blocking(move || {
-                                let world = state_clone.world.blocking_lock();
-                                let npc_mgr = state_clone.npc_manager.blocking_lock();
-                                mgr_clone.process_event(&evt_clone, &world, &npc_mgr)
-                            });
-                            match handle.await {
-                                Ok(Ok(())) => {}
-                                Ok(Err(e)) => tracing::warn!(error = %e, "character-log write failed"),
-                                Err(e) => tracing::warn!(error = %e, "character-log task panicked"),
-                            }
-                        }
-                        Err(tokio::sync::broadcast::error::RecvError::Lagged(_)) => continue,
-                        Err(tokio::sync::broadcast::error::RecvError::Closed) => break,
-                    },
-                }
-            }
-        }));
-    }
-
-    // ── Location-log subscriber ────────────────────────────────────────────
-    //
-    // Mirrors the character-log subscriber above; writes per-location
-    // markdown logs. Gated by the `location-logs` flag (default on).
-    {
-        let s = Arc::clone(&state);
-        let token = shutdown_token.clone();
-        handles.push(tokio::spawn(async move {
-            use parish_core::location_log::{FEATURE_FLAG, LocationLogManager};
-
-            let enabled = {
-                let cfg = s.config.lock().await;
-                !cfg.flags.is_disabled(FEATURE_FLAG)
-            };
-            if !enabled {
-                return;
-            }
-            let app_name = parish_core::game_mod::app_name_from_mod(&s.game_mod);
-            let mut current_branch = s.current_branch_id.lock().await.unwrap_or(1);
-            let mut manager = LocationLogManager::new(&app_name, current_branch, true);
-            let mut rx = {
-                let world = s.world.lock().await;
-                world.event_bus.subscribe()
-            };
-            {
-                let world = s.world.lock().await;
-                let npc_mgr = s.npc_manager.lock().await;
-                if let Err(e) = manager.write_all_profiles(&world, &npc_mgr) {
-                    tracing::warn!(error = %e, "location-log profile write failed");
-                }
-            }
-            loop {
-                tokio::select! {
-                    _ = token.cancelled() => break,
-                    result = rx.recv() => match result {
-                        Ok(event) => {
-                            // Rebind manager when the active branch has changed
-                            // (e.g. load_branch / create_branch). Mirrors the
-                            // character-log subscriber fix from #1011 (#1034).
-                            let bid = s.current_branch_id.lock().await.unwrap_or(1);
-                            if bid != current_branch {
-                                current_branch = bid;
-                                manager = LocationLogManager::new(&app_name, bid, true);
-                                let world = s.world.lock().await;
-                                let npc_mgr = s.npc_manager.lock().await;
-                                if let Err(e) = manager.write_all_profiles(&world, &npc_mgr) {
-                                    tracing::warn!(error = %e, "location-log profile write failed after branch switch");
-                                }
-                            }
-                            // Snapshot world and npc_manager state needed for name
-                            // resolution, then drop the locks before doing sync file I/O
-                            // to avoid blocking other tasks (#1012).
-                            // Clone the Arc<AppState> (cheap) and run the blocking
-                            // I/O task in a dedicated thread pool, avoiding lock
-                            // contention on the async side (#1012).
-                            let mgr_clone = manager.clone();
-                            let evt_clone = event.clone();
-                            let state_clone = Arc::clone(&s);
-                            let handle = tokio::task::spawn_blocking(move || {
-                                let world = state_clone.world.blocking_lock();
-                                let npc_mgr = state_clone.npc_manager.blocking_lock();
-                                mgr_clone.process_event(&evt_clone, &world, &npc_mgr)
-                            });
-                            match handle.await {
-                                Ok(Ok(())) => {}
-                                Ok(Err(e)) => tracing::warn!(error = %e, "location-log write failed"),
-                                Err(e) => tracing::warn!(error = %e, "location-log task panicked"),
-                            }
-                        }
-                        Err(tokio::sync::broadcast::error::RecvError::Lagged(_)) => continue,
-                        Err(tokio::sync::broadcast::error::RecvError::Closed) => break,
-                    },
-                }
-            }
-        }));
-    }
-
-    // ── Chat-transcript subscriber ─────────────────────────────────────────
-    //
-    // Writes the user-visible chat stream as JSONL paired with the inference
-    // log (for zippable bug reports), correlating NPC dialogue to the
-    // inference call via `parish.request_id`. The writer task + enable flag
-    // were created at session start on `AppState.chat_transcript_log`; this
-    // subscriber just forwards bus events to it. Per-session (not per-branch),
-    // so unlike the markdown log managers it never rebinds.
-    {
-        let s = Arc::clone(&state);
-        let token = shutdown_token.clone();
-        handles.push(tokio::spawn(async move {
-            // Always subscribe — even if logging starts disabled — so a
-            // mid-session `/inference-log on` is captured. `process_event`
-            // no-ops internally while the shared flag is off.
-            let mut rx = {
-                let world = s.world.lock().await;
-                world.event_bus.subscribe()
-            };
-            loop {
-                tokio::select! {
-                    _ = token.cancelled() => break,
-                    result = rx.recv() => match result {
-                        Ok(event) => {
-                            let world = s.world.lock().await;
-                            let npc_mgr = s.npc_manager.lock().await;
-                            s.chat_transcript_log.process_event(&event, &world, &npc_mgr);
-                        }
-                        Err(tokio::sync::broadcast::error::RecvError::Lagged(_)) => continue,
-                        Err(tokio::sync::broadcast::error::RecvError::Closed) => break,
-                    },
-                }
-            }
-        }));
-    }
-
-    // ── World tick (5 s) ─────────────────────────────────────────────────────
-    {
-        let s = Arc::clone(&state);
-        let token = shutdown_token.clone();
-        handles.push(tokio::spawn(async move {
-            // Round-robin cursor for budgeted gossip propagation (#466).
-            let mut gossip_cursor: usize = 0;
-            loop {
-                tokio::select! {
-                    _ = token.cancelled() => break,
-                    _ = tokio::time::sleep(Duration::from_secs(5)) => {}
-                }
-
-                {
-                    let world = s.world.lock().await;
-                    let npc_manager = s.npc_manager.lock().await;
-                    let mut snap = parish_core::ipc::snapshot_from_world(&world);
-                    snap.name_hints = parish_core::ipc::compute_name_hints(
-                        &world,
-                        &npc_manager,
-                        &s.pronunciations,
-                    );
-                    s.event_bus
-                        .emit_named(Topic::WorldUpdate, "world-update", &snap);
-                }
-
-                {
-                    // Snapshot the banshee flag outside the world/npc locks to avoid
-                    // nesting config → world, which inverts the project-wide
-                    // lock order.
-                    let banshee_enabled = {
-                        let cfg = s.config.lock().await;
-                        !cfg.flags.is_disabled("banshee")
-                    };
-
-                    let mut world = s.world.lock().await;
-                    let mut npc_mgr = s.npc_manager.lock().await;
-
-                    // Advance the world one pump through the single shared
-                    // helper (rule #12): weather + schedules + tier
-                    // reassignment + banshee + budgeted gossip. The server
-                    // intentionally leaves tier-4 to its own scheduling (it has
-                    // never dispatched tier-4 from this loop). The budgeted
-                    // gossip cursor round-robins across ticks (#466).
-                    {
-                        use parish_core::game_loop::{
-                            AdvanceOptions, GossipMode, WeatherMode, advance_world,
-                        };
-
-                        let mut rng = rand::rng();
-                        let report = advance_world(
-                            &mut world,
-                            &mut npc_mgr,
-                            &mut rng,
-                            AdvanceOptions {
-                                weather: WeatherMode::Single,
-                                run_banshee: banshee_enabled,
-                                gossip: GossipMode::Budgeted {
-                                    cursor: gossip_cursor,
-                                    budget: GOSSIP_BUDGET_PER_TICK,
-                                },
-                                run_tier4: false,
-                            },
-                        );
-                        gossip_cursor = report.gossip_cursor;
-                    }
-
-                    // Advance the generation counter so handle_game_input can
-                    // detect TOCTOU races (see issue #283).
-                    //
-                    // Skip while inference-paused: the player input is
-                    // mid-flight and the clock is frozen, so this tick is a
-                    // no-op from the player's perspective. Bumping the
-                    // counter anyway falsely tripped the TOCTOU guard.
-                    if !world.clock.is_inference_paused() {
-                        world.increment_tick_generation();
-                    }
-
-                    // #621 — Per-session tick metric. Emitted as a structured
-                    // tracing event so log-based metric tools can aggregate
-                    // tick counts per session without a Prometheus exporter.
-                    tracing::debug!(
-                        target: "parish_server::metrics",
-                        session_id = %s.session_id,
-                        tick = world.tick_generation,
-                        "session.tick"
-                    );
-                }
-            }
-        }));
-    }
-
-    // ── Inactivity tick (1 s) ────────────────────────────────────────────────
-    {
-        let s = Arc::clone(&state);
-        let token = shutdown_token.clone();
-        handles.push(tokio::spawn(async move {
-            loop {
-                tokio::select! {
-                    _ = token.cancelled() => break,
-                    _ = tokio::time::sleep(Duration::from_secs(1)) => {}
-                }
-                crate::routes::tick_inactivity(&s).await;
-            }
-        }));
-    }
-
-    // ── Autosave tick ────────────────────────────────────────────────────────
-    //
-    // #230 — Fixes: previously a fresh `Database::open` (and therefore a full
-    // `migrate()` round-trip) was executed on every tick.  Now we lazily open
-    // an `AsyncDatabase` the first time we have a save path and reuse it for
-    // all subsequent ticks.  All SQLite work is delegated to `spawn_blocking`
-    // inside `AsyncDatabase`, so a slow fsync can never stall the Tokio runtime.
-    {
-        let s = Arc::clone(&state);
-        let token = shutdown_token.clone();
-        handles.push(tokio::spawn(async move {
-            use parish_core::persistence::snapshot::GameSnapshot;
-            use parish_core::persistence::{AsyncDatabase, Database};
-            // Track whether the last autosave attempt failed so we only emit
-            // one user-visible warning per failure run, not one per tick.
-            let mut last_autosave_failed = false;
-            loop {
-                // Wait for the interval or cancellation.  Cancellation exits
-                // only after the *sleep* fires, so any in-flight autosave
-                // iteration completes before the task stops (#228).
-                tokio::select! {
-                    _ = token.cancelled() => break,
-                    _ = tokio::time::sleep(Duration::from_secs(AUTOSAVE_INTERVAL_SECS)) => {}
-                }
-
-                let save_path = s.save_path.lock().await.clone();
-                let branch_id = *s.current_branch_id.lock().await;
-
-                if let (Some(path), Some(bid)) = (save_path, branch_id) {
-                    // Snapshot the world state before touching the DB lock.
-                    let snapshot = {
-                        let world = s.world.lock().await;
-                        let npc_manager = s.npc_manager.lock().await;
-                        GameSnapshot::capture(&world, &npc_manager)
-                    };
-
-                    // Obtain (or open) the cached AsyncDatabase for this path.
-                    let db: Option<AsyncDatabase> = {
-                        let mut guard = s.save_db.lock().await;
-                        // If the cached path no longer matches the active save file
-                        // (e.g. after load-branch / new-save-file), discard the old handle.
-                        if guard.as_ref().is_some_and(|(p, _)| p != &path) {
-                            *guard = None;
-                        }
-                        if guard.is_none() {
-                            let path_clone = path.clone();
-                            match tokio::task::spawn_blocking(move || Database::open(&path_clone))
-                                .await
-                            {
-                                Ok(Ok(db)) => {
-                                    *guard = Some((path.clone(), AsyncDatabase::new(db)));
-                                }
-                                Ok(Err(e)) => {
-                                    tracing::warn!("Autosave: failed to open DB: {}", e);
-                                    if !last_autosave_failed {
-                                        s.event_bus.emit_named(
-                                            Topic::TextLog,
-                                            "text-log",
-                                            &parish_core::ipc::text_log(
-                                                "system",
-                                                "Autosave failed — could not open save file.",
-                                            ),
-                                        );
-                                        last_autosave_failed = true;
-                                    }
-                                    continue;
-                                }
-                                Err(e) => {
-                                    tracing::warn!("Autosave: spawn_blocking error: {}", e);
-                                    continue;
-                                }
-                            }
-                        }
-                        guard.as_ref().map(|(_, db)| db.clone())
-                    };
-
-                    if let Some(db) = db {
-                        match db.save_snapshot(bid, &snapshot).await {
-                            Ok(_) => {
-                                tracing::debug!("Session autosave complete");
-                                if last_autosave_failed {
-                                    s.event_bus.emit_named(
-                                        Topic::TextLog,
-                                        "text-log",
-                                        &parish_core::ipc::text_log(
-                                            "system",
-                                            "Autosave resumed successfully.",
-                                        ),
-                                    );
-                                }
-                                last_autosave_failed = false;
-                            }
-                            Err(e) => {
-                                tracing::warn!("Session autosave failed: {}", e);
-                                if !last_autosave_failed {
-                                    s.event_bus.emit_named(
-                                        Topic::TextLog,
-                                        "text-log",
-                                        &parish_core::ipc::text_log(
-                                            "system",
-                                            "Autosave failed — your progress may not be saved.",
-                                        ),
-                                    );
-                                    last_autosave_failed = true;
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }));
-    }
-
-    // ── Tier-2 simulation tick ────────────────────────────────────────────────
-    //
-    // Mirrors the Tier-2 polling task from `parish-tauri/src/setup.rs` and the
-    // `dispatch_headless_tier2_tick` helper from `parish-engine/src/headless.rs`.
-    // Previously missing from the web backend (TD-040), so `GossipSpread` events
-    // never fired on the Axum path. The shared `mint_tier2_gossip` helper (TD-030)
-    // is called for post-processing instead of repeating the loop body a third time
-    // (rule #12).
-    //
-    // Lock ordering: world → npc_manager (documented contract, matches the main
-    // world-tick above and the Tauri site — both hold world first, npc_manager
-    // second). The sim client and config are snapshotted before acquiring these
-    // locks so the LLM call is never made while holding a game-state lock.
-    {
-        let s = Arc::clone(&state);
-        let token = shutdown_token.clone();
-        handles.push(tokio::spawn(async move {
-            loop {
-                tokio::select! {
-                    _ = token.cancelled() => break,
-                    _ = tokio::time::sleep(Duration::from_secs(5)) => {}
-                }
-
-                // ── Check whether a Tier-2 tick is due ───────────────────────
-                let (needs_tick, in_flight, groups, time_desc, weather_str) = {
-                    let world = s.world.lock().await;
-                    let npc_mgr = s.npc_manager.lock().await;
-                    let now = world.clock.now();
-                    if !npc_mgr.needs_tier2_tick(now) || npc_mgr.tier2_in_flight() {
-                        continue;
-                    }
-                    use parish_core::npc::ticks::{Tier2Group, npc_snapshot_from_npc};
-                    let groups_map = npc_mgr.tier2_groups();
-                    if groups_map.is_empty() {
-                        continue;
-                    }
-                    let npc_names: std::collections::HashMap<_, _> =
-                        npc_mgr.all_npcs().map(|n| (n.id, n.name.clone())).collect();
-                    let groups: Vec<Tier2Group> = groups_map
-                        .into_iter()
-                        .filter_map(|(loc, npc_ids)| {
-                            let location_name = world
-                                .graph
-                                .get(loc)
-                                .map(|d| d.name.clone())
-                                .unwrap_or_else(|| format!("Location {}", loc.0));
-                            let npcs: Vec<_> = npc_ids
-                                .iter()
-                                .filter_map(|id| npc_mgr.get(*id))
-                                .map(|npc| npc_snapshot_from_npc(npc, &npc_names))
-                                .collect();
-                            if npcs.is_empty() {
-                                None
-                            } else {
-                                Some(Tier2Group {
-                                    location: loc,
-                                    location_name,
-                                    npcs,
-                                })
-                            }
-                        })
-                        .collect();
-                    if groups.is_empty() {
-                        continue;
-                    }
-                    let time_desc = world.clock.time_of_day().to_string();
-                    let weather_str = world.weather.to_string();
-                    (true, false, groups, time_desc, weather_str)
-                };
-                let _ = (needs_tick, in_flight); // consumed by the checks above
-
-                // Mark in-flight before releasing the lock so a concurrent tick
-                // cannot start a second batch.
-                {
-                    let mut npc_mgr = s.npc_manager.lock().await;
-                    npc_mgr.set_tier2_in_flight(true);
-                }
-
-                // ── Snapshot the sim client + model outside game-state locks ─
-                let (client_opt, model) = {
-                    use parish_core::config::InferenceCategory;
-                    let cfg = s.config.lock().await;
-                    let base_client = s.client.lock().await;
-                    cfg.resolve_category_client(InferenceCategory::Simulation, base_client.as_ref())
-                };
-
-                let Some(sim_client) = client_opt else {
-                    s.npc_manager.lock().await.set_tier2_in_flight(false);
-                    continue;
-                };
-
-                // ── Run one LLM call per Tier-2 group (outside locks) ────────
-                let lang = s.language_settings.clone();
-                let mut events = Vec::new();
-                for group in &groups {
-                    if let Some(evt) = parish_core::npc::ticks::run_tier2_for_group(
-                        &sim_client,
-                        &model,
-                        group,
-                        &time_desc,
-                        &weather_str,
-                        &lang,
-                        None,
-                    )
-                    .await
-                    {
-                        events.push(evt);
-                    }
-                }
-
-                // ── Apply events and mint gossip under game-state locks ───────
-                // Lock ordering: world → npc_manager.
-                {
-                    let mut world = s.world.lock().await;
-                    let mut npc_mgr = s.npc_manager.lock().await;
-                    let game_time = world.clock.now();
-
-                    let dbg = parish_core::game_loop::mint_tier2_gossip(
-                        &events,
-                        npc_mgr.npcs_mut(),
-                        game_time,
-                        &parish_core::config::NpcConfig::default(),
-                        &mut world,
-                    );
-                    npc_mgr.record_tier2_tick(game_time);
-                    npc_mgr.set_tier2_in_flight(false);
-
-                    let mut debug_events = s.debug_events.lock().await;
-                    if debug_events.len() >= crate::state::DEBUG_EVENT_CAPACITY {
-                        debug_events.pop_front();
-                    }
-                    debug_events.push_back(parish_core::debug_snapshot::DebugEvent {
-                        timestamp: String::new(),
-                        category: "tier2".to_string(),
-                        message: format!(
-                            "Tier 2 tick: {} events from {} groups{}",
-                            events.len(),
-                            groups.len(),
-                            if dbg.is_empty() {
-                                String::new()
-                            } else {
-                                format!("; {}", dbg.join(", "))
-                            },
-                        ),
-                    });
-                }
-            }
-        }));
-    }
-
-    handles
-}
-
-fn build_session_client(global: &GlobalState) -> (Option<AnyClient>, GameConfig) {
-    let config = global.template_config.clone();
-    let client = if config.provider_name == "simulator" {
-        Some(AnyClient::simulator())
-    } else if config.model_name.is_empty() && config.provider_name != "ollama" {
-        None
-    } else {
-        let provider_enum = parish_core::config::Provider::from_str_loose(&config.provider_name)
-            .unwrap_or_default();
-        Some(parish_core::inference::build_client(
-            &provider_enum,
-            &config.base_url,
-            config.api_key.as_deref(),
-            &global.inference_config, // (#417) use TOML-configured timeouts
-        ))
-    };
-    (client, config)
-}
-
-fn build_session_cloud_client(global: &GlobalState) -> Option<AnyClient> {
-    let config = &global.template_config;
-    config.cloud_api_key.as_deref().map(|key| {
-        let provider_enum = config
-            .cloud_provider_name
-            .as_deref()
-            .and_then(|p| parish_core::config::Provider::from_str_loose(p).ok())
-            .unwrap_or_else(|| {
-                parish_core::config::Provider::from_id("openrouter").unwrap_or_default()
-            });
-        parish_core::inference::build_client(
-            &provider_enum,
-            config
-                .cloud_base_url
-                .as_deref()
-                .unwrap_or("https://openrouter.ai/api"),
-            Some(key),
-            &global.inference_config, // (#417) use TOML-configured timeouts
-        )
-    })
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use parish_core::world::LocationId;
-
-    #[test]
-    fn autosave_interval_is_60_seconds() {
-        // Regression sensor: if this changes, update comment in session.rs and
-        // verify players won't lose more than AUTOSAVE_INTERVAL_SECS of progress.
-        assert_eq!(
-            AUTOSAVE_INTERVAL_SECS, 60,
-            "autosave interval changed — verify data-loss risk is acceptable"
-        );
-    }
-
-    /// Verifies that a fresh DB round-trips the Google OAuth link:
-    /// after `link_oauth`, both `find_by_oauth` and the identity store's
-    /// `get_account` return the stored values.
-    ///
-    /// This is the exact flow the callback + status endpoint use, so if
-    /// this test passes but the UI shows the user as signed out, the bug
-    /// is elsewhere (cookies, middleware, frontend).
-    #[test]
-    fn oauth_link_round_trips_on_fresh_db() {
-        let tmp = tempfile::tempdir().unwrap();
-        let reg = SessionRegistry::open(tmp.path()).unwrap();
-        reg.persist_new("sess_abc");
-        reg.link_oauth("google", "sub_123", "sess_abc", "John Doe");
-
-        assert_eq!(
-            reg.find_by_oauth("google", "sub_123"),
-            Some("sess_abc".to_string()),
-            "find_by_oauth should return the linked session_id"
-        );
-
-        let conn = crate::session_store_impl::open_sessions_db(tmp.path()).unwrap();
-        let store = crate::session_store_impl::SqliteIdentityStore::new(conn);
-        assert_eq!(
-            store.get_account("sess_abc"),
-            Some(("sub_123".to_string(), "John Doe".to_string())),
-            "get_account should return (sub, display_name)"
-        );
-    }
-
-    /// Verifies the migration from a pre-display_name schema to the
-    /// current schema: opening a DB that was created with the old schema
-    /// should add the `display_name` column, and subsequent link_oauth
-    /// + identity-store `get_account` calls should work end-to-end.
-    #[test]
-    fn oauth_link_round_trips_on_migrated_db() {
-        let tmp = tempfile::tempdir().unwrap();
-        let db_path = tmp.path().join("sessions.db");
-
-        // Simulate an existing DB created with the pre-display_name schema.
-        {
-            let conn = rusqlite::Connection::open(&db_path).unwrap();
-            conn.execute_batch(
-                "CREATE TABLE sessions (
-                    id TEXT PRIMARY KEY,
-                    created_at TEXT NOT NULL,
-                    last_active TEXT NOT NULL
-                );
-                CREATE TABLE oauth_accounts (
-                    provider         TEXT NOT NULL,
-                    provider_user_id TEXT NOT NULL,
-                    session_id       TEXT NOT NULL,
-                    PRIMARY KEY (provider, provider_user_id)
-                );",
-            )
-            .unwrap();
-            // Insert a row that predates the display_name column.
-            conn.execute(
-                "INSERT INTO oauth_accounts (provider, provider_user_id, session_id) \
-                 VALUES ('google', 'legacy_sub', 'legacy_sess')",
-                [],
-            )
-            .unwrap();
-        }
-
-        // Re-open through SessionRegistry — this should ADD COLUMN display_name.
-        let reg = SessionRegistry::open(tmp.path()).unwrap();
-
-        let conn = crate::session_store_impl::open_sessions_db(tmp.path()).unwrap();
-        let store = crate::session_store_impl::SqliteIdentityStore::new(conn);
-
-        // Legacy row has empty display_name (default).
-        assert_eq!(
-            store.get_account("legacy_sess"),
-            Some(("legacy_sub".to_string(), String::new())),
-        );
-
-        // New link writes the display_name column correctly.
-        reg.persist_new("sess_new");
-        reg.link_oauth("google", "sub_new", "sess_new", "Jane Doe");
-        assert_eq!(
-            store.get_account("sess_new"),
-            Some(("sub_new".to_string(), "Jane Doe".to_string())),
-        );
-    }
-
-    // ── #466 gossip budget round-robin ──────────────────────────────────────
-    //
-    // The budgeting math (`budgeted_round_robin`) and its unit tests moved to
-    // `parish_core::game_loop::world_pump` in #1159 — the server now drives the
-    // single shared `advance_world` pump with `GossipMode::Budgeted`.
-
-    // ── #482 disk-session purge ─────────────────────────────────────────────
-
-    /// Overwrites sessions.id's last_active to a fixed ISO timestamp so
-    /// tests can pin "how idle" a row is without sleeping through the
-    /// real retention window.
-    fn backdate_session(reg: &SessionRegistry, session_id: &str, last_active_iso: &str) {
-        let db = reg.db.lock().unwrap();
-        db.execute(
-            "UPDATE sessions SET last_active = ?1 WHERE id = ?2",
-            rusqlite::params![last_active_iso, session_id],
-        )
-        .unwrap();
-    }
-
-    #[test]
-    fn purge_expired_removes_old_row_and_save_dir() {
-        // Use a valid UUID v4 format ID — the #595 path-traversal guard
-        // requires session IDs to be hex+hyphen only (matching UUID v4).
-        let expired_id = "e1111111-1111-4111-a111-111111111111";
-        let tmp = tempfile::tempdir().unwrap();
-        let reg = SessionRegistry::open(tmp.path()).unwrap();
-        reg.persist_new(expired_id);
-        // Fresh row + fake saves/<id>/ directory.
-        let save_dir = tmp.path().join(expired_id);
-        std::fs::create_dir_all(&save_dir).unwrap();
-        std::fs::write(save_dir.join("parish_001.db"), b"fake").unwrap();
-        // Backdate to 90 days ago so any reasonable retention sweep
-        // picks it up.
-        let old = (chrono::Utc::now() - chrono::Duration::days(90)).to_rfc3339();
-        backdate_session(&reg, expired_id, &old);
-
-        let purged = reg.purge_expired_disk_sessions(tmp.path(), Duration::from_secs(30 * 86_400));
-        assert_eq!(purged, 1);
-        assert!(!reg.exists_in_db(expired_id));
-        assert!(
-            !save_dir.exists(),
-            "saves directory must be deleted after purge"
-        );
-    }
-
-    #[test]
-    fn purge_expired_preserves_recent_sessions() {
-        // Use a valid UUID v4 format ID — the #595 path-traversal guard
-        // requires session IDs to be hex+hyphen only (matching UUID v4).
-        let recent_id = "ece11111-1111-4111-a111-111111111111";
-        let tmp = tempfile::tempdir().unwrap();
-        let reg = SessionRegistry::open(tmp.path()).unwrap();
-        reg.persist_new(recent_id);
-        let save_dir = tmp.path().join(recent_id);
-        std::fs::create_dir_all(&save_dir).unwrap();
-
-        // last_active set to `now` by persist_new — well inside the
-        // 30-day retention window.
-        let purged = reg.purge_expired_disk_sessions(tmp.path(), Duration::from_secs(30 * 86_400));
-        assert_eq!(purged, 0);
-        assert!(reg.exists_in_db(recent_id));
-        assert!(save_dir.exists());
-    }
-
-    #[test]
-    fn purge_expired_drops_linked_oauth_rows() {
-        // Use a valid UUID v4 format ID — the #595 path-traversal guard
-        // requires session IDs to be hex+hyphen only (matching UUID v4).
-        let expired_linked_id = "e1111111-1111-4111-a111-111111111112";
-        let tmp = tempfile::tempdir().unwrap();
-        let reg = SessionRegistry::open(tmp.path()).unwrap();
-        reg.persist_new(expired_linked_id);
-        reg.link_oauth("google", "sub_legacy", expired_linked_id, "Old User");
-        let old = (chrono::Utc::now() - chrono::Duration::days(90)).to_rfc3339();
-        backdate_session(&reg, expired_linked_id, &old);
-
-        let purged = reg.purge_expired_disk_sessions(tmp.path(), Duration::from_secs(30 * 86_400));
-        assert_eq!(purged, 1);
-        // The OAuth link is gone too — otherwise a fresh login for
-        // `sub_legacy` would resurrect a dead session_id with no DB row.
-        assert_eq!(reg.find_by_oauth("google", "sub_legacy"), None);
-    }
-
-    #[test]
-    fn purge_expired_handles_missing_save_dir_gracefully() {
-        // Use a valid UUID v4 format ID — the #595 path-traversal guard
-        // requires session IDs to be hex+hyphen only (matching UUID v4).
-        let ghost_id = "abb51111-1111-4111-a111-111111111111";
-        let tmp = tempfile::tempdir().unwrap();
-        let reg = SessionRegistry::open(tmp.path()).unwrap();
-        reg.persist_new(ghost_id);
-        // No saves/<id>/ directory was ever created. Purge must still
-        // delete the DB row and return 1 — filesystem absence is fine.
-        let old = (chrono::Utc::now() - chrono::Duration::days(90)).to_rfc3339();
-        backdate_session(&reg, ghost_id, &old);
-
-        let purged = reg.purge_expired_disk_sessions(tmp.path(), Duration::from_secs(30 * 86_400));
-        assert_eq!(purged, 1);
-        assert!(!reg.exists_in_db(ghost_id));
-    }
-
-    // ── #595 path traversal guard ────────────────────────────────────────────
-
-    /// A session ID containing `..` must not cause `remove_dir_all` to
-    /// operate outside the saves root.  The traversal ID is rejected before
-    /// the filesystem is touched; a sibling directory must survive intact.
-    #[test]
-    fn purge_expired_rejects_path_traversal_id() {
-        let outer = tempfile::tempdir().unwrap();
-        // The "saves root" lives one level below outer so there is a parent
-        // directory to try to traverse into.
-        let saves_root = outer.path().join("saves");
-        std::fs::create_dir_all(&saves_root).unwrap();
-
-        // A sibling directory that a traversal payload would try to delete.
-        let sibling = outer.path().join("sensitive");
-        std::fs::create_dir_all(&sibling).unwrap();
-        std::fs::write(sibling.join("secret.txt"), b"do not delete").unwrap();
-
-        // Set up a SessionRegistry using saves_root as the root.
-        let reg = SessionRegistry::open(&saves_root).unwrap();
-
-        // Directly insert a row with a traversal ID (bypassing the normal
-        // UUID generation path to simulate a tampered/corrupted DB).
-        {
-            let db = reg.db.lock().unwrap();
-            db.execute(
-                "INSERT INTO sessions (id, created_at, last_active) VALUES (?1, ?2, ?2)",
-                rusqlite::params!["../sensitive", "2000-01-01T00:00:00Z"],
-            )
-            .unwrap();
-        }
-
-        // Create a fake directory at saves_root/../sensitive to give
-        // remove_dir_all something to hit if the guard fails.
-        // (sibling already exists above — that's the target.)
-
-        let purged = reg
-            .purge_expired_disk_sessions(&saves_root, Duration::from_secs(0 /* always expired */));
-
-        // The DB row is deleted (purge still counts it).
-        assert_eq!(purged, 1);
-        // The sibling directory must NOT have been removed.
-        assert!(
-            sibling.exists(),
-            "path traversal guard must prevent deletion of directories outside saves root"
-        );
-        assert!(
-            sibling.join("secret.txt").exists(),
-            "sensitive file must survive"
-        );
-    }
-
-    /// IDs with non-hex/non-hyphen characters (including `/` and `\`) are
-    /// rejected by the allowlist even if they don't look like `..` traversals.
-    #[test]
-    fn purge_expired_rejects_ids_with_unsafe_characters() {
-        let tmp = tempfile::tempdir().unwrap();
-        let reg = SessionRegistry::open(tmp.path()).unwrap();
-
-        // Directly insert rows with unsafe IDs.
-        let unsafe_ids = [
-            "../../etc/passwd",
-            "foo/bar",
-            "foo\\bar",
-            "abc def",
-            "abc\0def",
-        ];
-        {
-            let db = reg.db.lock().unwrap();
-            for id in &unsafe_ids {
-                db.execute(
-                    "INSERT INTO sessions (id, created_at, last_active) VALUES (?1, ?2, ?2)",
-                    rusqlite::params![id, "2000-01-01T00:00:00Z"],
-                )
-                .unwrap();
-            }
-        }
-
-        // None of these should cause a panic or an out-of-root deletion.
-        let purged = reg.purge_expired_disk_sessions(tmp.path(), Duration::from_secs(0));
-        // All rows are deleted from the DB.
-        assert_eq!(purged, unsafe_ids.len());
-        // The saves root itself is intact.
-        assert!(tmp.path().exists(), "saves root must still exist");
-    }
-
-    /// A well-formed UUID session ID must still be cleaned up normally —
-    /// the path-traversal guard must not break the happy path.
-    #[test]
-    fn purge_expired_uuid_id_still_cleaned_up() {
-        let tmp = tempfile::tempdir().unwrap();
-        let reg = SessionRegistry::open(tmp.path()).unwrap();
-        let id = "a1b2c3d4-e5f6-4789-abcd-ef0123456789";
-        reg.persist_new(id);
-        let save_dir = tmp.path().join(id);
-        std::fs::create_dir_all(&save_dir).unwrap();
-
-        let old = (chrono::Utc::now() - chrono::Duration::days(90)).to_rfc3339();
-        backdate_session(&reg, id, &old);
-
-        let purged = reg.purge_expired_disk_sessions(tmp.path(), Duration::from_secs(30 * 86_400));
-        assert_eq!(purged, 1);
-        assert!(!save_dir.exists(), "save directory must be removed");
-    }
-
-    /// Regression test for #230: the autosave path must reuse a single
-    /// `AsyncDatabase` across multiple saves rather than reopening the file
-    /// (and re-running `migrate()`) on every tick.
-    ///
-    /// Verifies:
-    /// 1. Opening the DB once and calling `save_snapshot` N times produces N
-    ///    snapshots in the database (i.e. the handle is reused, not replaced).
-    /// 2. The snapshot count matches the number of save calls — if a new
-    ///    connection were opened each time, the per-call migrate() would not
-    ///    duplicate rows, but we confirm the handle is indeed reused by checking
-    ///    that the Arc inside AsyncDatabase stays alive across calls.
-    #[tokio::test]
-    async fn autosave_reuses_async_database_across_ticks() {
-        use parish_core::persistence::snapshot::{ClockSnapshot, GameSnapshot};
-        use parish_core::persistence::{AsyncDatabase, Database};
-
-        let tmp = tempfile::NamedTempFile::new().unwrap();
-
-        // Open the DB once — exactly what the fixed autosave tick does.
-        let db = Database::open(tmp.path()).unwrap();
-        let async_db = AsyncDatabase::new(db);
-
-        let branch = async_db.find_branch("main").await.unwrap().unwrap();
-
-        fn make_snapshot() -> GameSnapshot {
-            GameSnapshot {
-                player_location: LocationId(1),
-                weather: "Clear".to_string(),
-                text_log: vec![],
-                clock: ClockSnapshot {
-                    game_time: chrono::DateTime::<chrono::Utc>::from_timestamp(0, 0).unwrap(),
-                    speed_factor: 36.0,
-                    paused: false,
-                },
-                npcs: vec![],
-                last_tier2_game_time: None,
-                last_tier3_game_time: None,
-                last_tier4_game_time: None,
-                introduced_npcs: Default::default(),
-                visited_locations: std::collections::HashSet::new(),
-                visited_order: Vec::new(),
-                edge_traversals: Default::default(),
-                gossip_network: Default::default(),
-                conversation_log: Default::default(),
-                player_name: None,
-                npcs_who_know_player_name: Default::default(),
-            }
-        }
-
-        // Simulate three autosave ticks using the same handle.
-        for _ in 0..3 {
-            async_db
-                .save_snapshot(branch.id, &make_snapshot())
-                .await
-                .expect("autosave tick should succeed with reused connection");
-        }
-
-        // All three snapshots must be present; branch_log returns most-recent-first.
-        let log = async_db.branch_log(branch.id).await.unwrap();
-        assert_eq!(
-            log.len(),
-            3,
-            "three autosave ticks via the same AsyncDatabase must produce three snapshots"
-        );
-    }
-
-    // ── Admission control unit tests (#620) ──────────────────────────────────
-
-    /// `active_count` reflects the current in-memory session count.
-    #[test]
-    fn active_count_reflects_in_memory_sessions() {
-        let tmp = tempfile::tempdir().unwrap();
-        let reg = SessionRegistry::open(tmp.path()).unwrap();
-        assert_eq!(reg.active_count(), 0);
-
-        // Insert a dummy entry — Arc::new with a dummy SessionEntry requires
-        // fields we can't easily construct here, so just verify the count at 0.
-        // The is_at_capacity / active_count pairing is tested in the next test.
-        assert_eq!(
-            reg.active_count(),
-            0,
-            "fresh registry must report 0 sessions"
-        );
-    }
-
-    /// `is_at_capacity` returns false when under the cap and true at/above it.
-    #[test]
-    fn is_at_capacity_returns_false_below_cap_true_at_cap() {
-        let tmp = tempfile::tempdir().unwrap();
-        let reg = SessionRegistry::open(tmp.path()).unwrap();
-
-        // With 0 sessions and cap=1, not at capacity.
-        assert!(
-            !reg.is_at_capacity(1),
-            "0 sessions, cap=1: should not be at capacity"
-        );
-        // rejection_count must not have incremented.
-        assert_eq!(reg.rejection_count.load(Ordering::Relaxed), 0);
-
-        // Artificially reach the cap by checking is_at_capacity(0): cap=0 means
-        // every new attempt is rejected.
-        assert!(
-            reg.is_at_capacity(0),
-            "0 sessions, cap=0: should be at capacity"
-        );
-        assert_eq!(
-            reg.rejection_count.load(Ordering::Relaxed),
-            1,
-            "rejection_count must increment"
-        );
-
-        // A second rejection increments again.
-        assert!(reg.is_at_capacity(0));
-        assert_eq!(reg.rejection_count.load(Ordering::Relaxed), 2);
-    }
-
-    /// `rejection_count` is not incremented for calls that are under the cap.
-    #[test]
-    fn rejection_count_not_incremented_below_cap() {
-        let tmp = tempfile::tempdir().unwrap();
-        let reg = SessionRegistry::open(tmp.path()).unwrap();
-        // active_count() == 0 < cap=10 → no rejection
-        let _ = reg.is_at_capacity(10);
-        let _ = reg.is_at_capacity(10);
-        assert_eq!(reg.rejection_count.load(Ordering::Relaxed), 0);
-    }
-}
+pub(crate) const GOSSIP_BUDGET_PER_TICK: usize = 20;
